blog/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==0.10.1

# Flask Packages
Flask-Migrate==1.4.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-User==0.6.9
Flask-WTF==0.12
Flask-Login==0.2.11
Flask-Moment=0.5.1

# Development tools
pytest==2.7.2
pytest-flask==0.8.1
pytest-cov==1.8.1

